requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
robotframework==4.1
robotframework-pythonlibcore==3.0.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==5.1.3
selenium==3.141.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.6
webdriver-manager==4.0.0
wsproto==1.2.0
